+++ AegeanTools/BANE.py
@@ -167,7 +167,8 @@ def sigma_filter(filename, region, step_size, box_size, shape, domask, sid):
     # Figure out how many axes are in the datafile
     NAXIS = fits.getheader(filename)["NAXIS"]
 
-    with fits.open(filename, memmap=True) as a:
+    # For some reason we can't memmap a file with BSCALE not 1.0, so we signore it now and scale it later
+    with fits.open(filename, memmap=True, do_not_scale_image_data=True) as a:
         if NAXIS == 2:
             data = a[0].section[data_row_min:data_row_max, 0:shape[1]]
         elif NAXIS == 3:
@@ -178,6 +179,11 @@ def sigma_filter(filename, region, step_size, box_size, shape, domask, sid):
             logging.error("Too many NAXIS for me {0}".format(NAXIS))
             logging.error("fix your file to be more sane")
             raise Exception("Too many NAXIS")
+
+    # Manually scale the data if BSCALE is not 1.0
+    header = fits.getheader(filename)
+    if 'BSCALE' in header:
+        data *= header['BSCALE']
 
     row_len = shape[1]
 
@@ -436,18 +442,23 @@ def filter_image(im_name, out_base, step_size=None, box_size=None, twopass=False
         bkg_out = '_'.join([os.path.expanduser(out_base), 'bkg.fits'])
         rms_out = '_'.join([os.path.expanduser(out_base), 'rms.fits'])
 
+        # Test for BSCALE and scale back if needed before we write to a file
+        bscale = 1.0
+        if 'BSCALE' in header:
+            bscale = header['BSCALE']
+
         # compress
         if compressed:
-            hdu = fits.PrimaryHDU(bkg)
+            hdu = fits.PrimaryHDU(bkg/bscale)
             hdu.header = copy.deepcopy(header)
             hdulist = fits.HDUList([hdu])
             compress(hdulist, step_size[0], bkg_out)
             hdulist[0].header = copy.deepcopy(header)
-            hdulist[0].data = rms
+            hdulist[0].data = rms/bscale
             compress(hdulist, step_size[0], rms_out)
         else:
-            write_fits(bkg, header, bkg_out)
-            write_fits(rms, header, rms_out)
+            write_fits(bkg/bscale, header, bkg_out)
+            write_fits(rms/bscale, header, rms_out)
 
     return bkg, rms